python-backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.4",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "langchain>=0.3,<0.4",
    "langchain-core>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "numpy>=1.26",
    "openai>=2.8.0",
    "pdfplumber>=0.11.8",
    "pinecone>=7.3.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
